dart: rename enums.fbs
