+++ tests/DartTest.sh
@@ -16,7 +16,10 @@
 
 pushd "$(dirname $0)" >/dev/null
 
-command -v dart >/dev/null 2>&1 || { echo >&2 "Dart tests require dart to be in path but it's not installed.  Aborting."; exit 1; }
+command -v dart >/dev/null 2>&1 || {
+    echo >&2 "Dart tests require dart to be in path but it's not installed.  Aborting."
+    exit 1
+}
 # output required files to the dart folder so that pub will be able to
 # distribute them and more people can more easily run the dart tests
 ../flatc --dart --gen-object-api -I include_test -o ../dart/test monster_test.fbs
@@ -28,7 +31,7 @@ cp monster_test.fbs ../dart/test
 
 cd ../dart
 
-../flatc --dart --gen-object-api -o ./test ./test/list_of_enums.fbs
+../flatc --dart --gen-object-api -o ./test ./test/enums.fbs
 ../flatc --dart --gen-object-api -o ./test ./test/bool_structs.fbs
 
 # update packages
